src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "pillow>=12.0.0",
    "pymupdf>=1.26.5",
    "pystray>=0.19.5",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "pyinstaller>=6.9.0",
    "watchdog>=2.1.0",
    "pyperclip>=1.8.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.21",
    "markdown>=3.10",
]
